requirements.txt
Flask~=2.3.2
numpy~=1.24.3
opencv_python~=4.5.5.62
Pillow~=9.0.1
python-dotenv~=1.0.0
torch~=1.13.1
gdown~=4.7.1
torchvision~=0.14.1
scipy~=1.10.0